Adding double support did not fix numerical precision errors.

+++ k2build.cl
@@ -277,7 +277,7 @@ __kernel void  k2(
 #define union(a,b) T_min(a,b)
 #define intersection(a,b) T_max(a,b)
 #define subtraction(a,b) T_max(a,-b)
-#define Vector3f(x,y,z) ((float3)((float)(x),(float)(y),(float)(z)))
+#define Vector3f(x,y,z) ((double3)((float)(x),(float)(y),(float)(z)))
 #define signOfInt(i) (i>0?1:(i<0?-1:(0)))
 
 #define DIRECTION_X 0
@@ -350,12 +350,12 @@ __kernel void  k2(
 	}
 
 
-	float maxComponent(float3 v){
+	float maxComponent(double3 v){
 
 		return T_max(v.x,T_max(v.y,v.z));
 	}
 
-	float box(float3 point, float3 center, float3 halfDiameter ){
+	float box(double3 point, double3 center, double3 halfDiameter ){
 
 		point=fabs(point-center);
 
@@ -363,7 +363,7 @@ __kernel void  k2(
 		
 	}
 
-	float getComponent(float3 v, int component){
+	float getComponent(double3 v, int component){
 
 		if(component==0) return v.x;
 		if(component==1) return v.y;
@@ -373,19 +373,19 @@ __kernel void  k2(
 	
 	}
 
-	float3 termProduct(float3 a, float3 b){
+	double3 termProduct(double3 a, double3 b){
 
 		return Vector3f(a.x*b.x,a.y*b.y,a.z*b.z);
 	}
 
-	float3 swizzle(float3 v, int a, int b, int c){
+	double3 swizzle(double3 v, int a, int b, int c){
 		return Vector3f(getComponent(v,a),getComponent(v,b),getComponent(v,c));
 	}
 
 
 
 	//Messed up the orientation at first
-	float _hilbertUnitCell(float3 v){
+	float _hilbertUnitCell(double3 v){
 
 		float d1 = box(v,Vector3f(-0.5,-0.5,0.0),Vector3f(lineWidth,lineWidth,0.5+lineWidth));
 		float d2 = box(v,Vector3f(0.5,-0.5,0.0),Vector3f(lineWidth,lineWidth,0.5+lineWidth));
@@ -414,7 +414,7 @@ __kernel void  k2(
 	}
 
 	//Messed up the orientation at first
-	float hilbertUnitCell(float3 v){
+	float hilbertUnitCell(double3 v){
 
 		v=termProduct(swizzle(v,1,0,2),Vector3f(1,-1,1));
 		v=termProduct(swizzle(v,2,1,0),Vector3f(1,1,-1));
@@ -423,9 +423,9 @@ __kernel void  k2(
 	}
 
 
-	float putHilbert(float3 v,int x, int y, int z)
+	float putHilbert(double3 v,int x, int y, int z)
 	{
-		float3 c = Vector3f(x/3.0,y/3.0,z/3.0);
+		double3 c = Vector3f(x/3.0,y/3.0,z/3.0);
 		v=Vector3f(v.x-c.x,v.y-c.y,v.z-c.z);
 		v=Vector3f(3.0*v.x,3.0*v.y,3.0*v.z);
 
@@ -446,9 +446,9 @@ __kernel void  k2(
 		float m21=quadrantMatrices[matrixOffset+7];
 		float m22=quadrantMatrices[matrixOffset+8];
 
-		float3 mc0 = Vector3f(m00,m01,m02);
-		float3 mc1 = Vector3f(m10,m11,m12);
-		float3 mc2 = Vector3f(m20,m21,m22); 
+		double3 mc0 = Vector3f(m00,m01,m02);
+		double3 mc1 = Vector3f(m10,m11,m12);
+		double3 mc2 = Vector3f(m20,m21,m22); 
 
 		float A = dot(v,mc0);
 		float B = dot(v,mc1);
@@ -458,7 +458,7 @@ __kernel void  k2(
 
 	}
 
-	float putShaft(float3 v, float3 center, float halfWidth, float halfLength, int direction){
+	float putShaft(double3 v, double3 center, float halfWidth, float halfLength, int direction){
 		float d = MAX_DISTANCE;
 		switch(direction){
 			case DIRECTION_X:
@@ -478,9 +478,9 @@ __kernel void  k2(
 		return d;
 	}
 
-	float putConnector(float3 v, int largeI, int largeJ, int largeK, int i, int j, int k, int direction){
+	float putConnector(double3 v, int largeI, int largeJ, int largeK, int i, int j, int k, int direction){
 
-		float3 center = Vector3f(
+		double3 center = Vector3f(
 
 (largeI*1.0+i/2.0)*1/3.0,
 (largeJ*1.0+j/2.0)*1/3.0,
@@ -492,7 +492,7 @@ __kernel void  k2(
 
 	}
 
-	float putConnectors(float3 v){
+	float putConnectors(double3 v){
 
 		float d = MAX_DISTANCE;
 		d=union(d,putConnector(v,0,-1,1,0,1,1,DIRECTION_X));
@@ -511,7 +511,7 @@ __kernel void  k2(
 	}
 
 
-	float hilbert_sdf(float3 v){
+	float hilbert_sdf(double3 v){
 
 
 		//return hilbertUnitCell(v);
@@ -535,21 +535,21 @@ __kernel void  k2(
 
 
         
-        float sd0( float3 v){
+        float sd0( double3 v){
 
             return MAX_DISTANCE;
 
         }
         
 
-        float sd1( float3 v){
+        float sd1( double3 v){
 
             return length(v)-0.5;
 
         }
         
 
-        float sd2( float3 v){
+        float sd2( double3 v){
 
             
 
@@ -563,7 +563,7 @@ __kernel void  k2(
         }
         
 
-        float sd3( float3 v){
+        float sd3( double3 v){
 
             
     v=fabs(v);
@@ -573,14 +573,14 @@ __kernel void  k2(
         }
         
 
-        float sd4( float3 v){
+        float sd4( double3 v){
 
             return length(v)-0.5;
 
         }
         
 
-        float sd5( float3 v){
+        float sd5( double3 v){
 
             
 
@@ -594,7 +594,7 @@ __kernel void  k2(
         }
         
 
-        float sd6( float3 v){
+        float sd6( double3 v){
 
             
     v=fabs(v);
@@ -604,7 +604,7 @@ __kernel void  k2(
         }
         
 
-        float sd7( float3 v){
+        float sd7( double3 v){
 
              
 
@@ -615,7 +615,7 @@ __kernel void  k2(
         }
         
 
-        float sd8( float3 v){
+        float sd8( double3 v){
 
             
 
@@ -639,20 +639,20 @@ __kernel void  k2(
         
 
         
-        float3 shader0 (float3 gv, float3 lv, float3 n){
+        double3 shader0 (double3 gv, double3 lv, double3 n){
 
             return fabs(n);
 
         }
         
 
-        float3 shader1 (float3 gv, float3 lv, float3 n){
+        double3 shader1 (double3 gv, double3 lv, double3 n){
 
             
         
-        float3 n_g = n.x*rgt_g+n.y*upp_g+n.z*fwd_g;
+        double3 n_g = n.x*rgt_g+n.y*upp_g+n.z*fwd_g;
 
-        float L = dot(n_g,(float3)(0.0,0.0,-1.0)); return (float3)(L,L,L);
+        float L = dot(n_g,(double3)(0.0,0.0,-1.0)); return (double3)(L,L,L);
 
 
 
@@ -662,7 +662,7 @@ __kernel void  k2(
         
 
 
-        float sdf_bank(float3 v, unsigned char shape_id){
+        float sdf_bank(double3 v, unsigned char shape_id){
 
             switch(shape_id){
 
@@ -700,7 +700,7 @@ case 8: return sd8(v); break;
 
         }
 
-        float3 shader_bank(float3 gv, float3 lv, float3 n, unsigned char material_id){
+        double3 shader_bank(double3 gv, double3 lv, double3 n, unsigned char material_id){
 
 
             switch(material_id){
@@ -714,7 +714,7 @@ case 1: return shader1(gv,lv,n); break;
 
             }
 
-            return (float3)(1.0, 1.0, 1.0);
+            return (double3)(1.0, 1.0, 1.0);
         }
